adding arch linux support

+++ funcs.sh
@@ -24,7 +24,7 @@ chkroot(){
 }
 
 chknorm(){
-	if [ "$(id -u)" = "0" ] && ! $JL_debian; then
+	if [ "$(id -u)" = "0" ]; then
 	  err_out "you need to run this as a normal user, not root."
 	  exit 1
 	fi
@@ -32,9 +32,13 @@ chknorm(){
 
 mode_select(){
 	PS3='Please select a mode (#?): '
-	opts="Ubuntu Debian"
+	opts="Ubuntu Debian ArchLinux"
 	select opt in $opts; do #must not double quote
 		case $opt in
+		    ArchLinux)
+		        echo archlinux >/dev/stdout
+		        break
+		        ;;
 			Ubuntu)
 				echo ubuntu >/dev/stdout
 				break
@@ -182,10 +186,13 @@ refresh_network(){
 	  	exit 1
 	fi
 	cd "$livedir"
-	msg_out "Preparing network connection for chrootin $livedir..."
-	cp /etc/hosts edit/etc/
-	rm edit/etc/resolv.conf
-	cp -L /etc/resolv.conf edit/etc/
+	msg_out "Preparing network connection for chroot in $livedir..."
+    #~ if ! $JL_archlinux; then
+        #~ cp /etc/hosts edit/etc/hosts
+    #~ fi
+    cp /etc/hosts edit/etc/
+    rm -f edit/etc/resolv.conf || err_out "Could not remove resolv.conf, run 'JLopt -rn' if network isn't available"
+    cp -L /etc/resolv.conf edit/etc/resolv.conf
 	msg_out "Network connection shlould be available in chroot now....."
 }
 
@@ -375,24 +382,30 @@ umount_fs(){
 	    edit="$livedir/edit/"
 	fi
     edit=$(sanit_path "$edit")/ #must have / at end
-	if mount |awk '{print $3}' |grep -qF "${edit}"proc; then
+    mounted=$(mount |awk '{print $3}')
+	if echo "$mounted" |grep -qF "${edit}"proc; then
 		if umount "${edit}"proc || umount -lf "${edit}"proc ; then
 			msg_out "unmount proc success"
 		fi
 	fi
-	if mount |awk '{print $3}' |grep -qF "${edit}"sys; then
+	if echo "$mounted" |grep -qF "${edit}"sys; then
 		if umount "${edit}"sys || umount -lf "${edit}"sys ; then
 			msg_out "unmount sys success"
 		fi
 	fi
-	if mount |awk '{print $3}' |grep -qF "${edit}"dev/pts; then
+	if echo "$mounted" |grep -qF "${edit}"dev/pts; then
 		if umount "${edit}"dev/pts || umount -lf "${edit}"dev/pts ; then
 			msg_out "unmount devpts success"
 		fi
 	fi
-	if mount |awk '{print $3}' |grep -qF "${edit}"dev; then
+	if echo "$mounted" |grep -qF "${edit}"dev; then
 		if umount "${edit}"dev || umount -lf "${edit}"dev; then
 			msg_out "unmount dev success"
+		fi
+	fi
+	if echo "$mounted" |grep -qF "${livedir}/"mnt; then
+		if umount "${livedir}/"mnt || umount -lf "${livedir}/"mnt; then
+			msg_out "unmount mnt success"
 		fi
 	fi
 	remove_fsentry_fstab "$edit"
@@ -447,6 +460,19 @@ rebuild_initrd(){
 	# fi
 }
 
+rebuild_initramfs(){
+    $CHROOT pacman -Syu --force archiso linux
+    HOOKS='"base udev memdisk archiso_shutdown archiso archiso_loop_mnt archiso_pxe_common archiso_pxe_nbd archiso_pxe_http archiso_pxe_nfs archiso_kms block pcmcia filesystems keyboard"'
+    sed -i.bak "s/HOOKS=\"[^\"]*\"/HOOKS=$HOOKS/" edit/etc/mkinitcpio.conf
+    $CHROOT mkinitcpio -p linux
+    update_cp edit/boot/vmlinuz-linux "extracted/arch/boot/$JL_arch/vmlinuz"
+    update_cp edit/boot/initramfs-linux.img "extracted/arch/boot/$JL_arch/archiso.img"
+    #update efi 
+    mount -t vfat -o loop extracted/EFI/archiso/efiboot.img mnt
+    cp extracted/arch/boot/$JL_arch/vmlinuz mnt/EFI/archiso/vmlinuz.efi || exit 1
+    cp extracted/arch/boot/x86_64/archiso.img mnt/EFI/archiso/archiso.img || exit 1
+}
+
 jl_clean(){
 	kerver=$(uname -r)
 	cd "$livedir" #exported from jlcd_start
@@ -460,7 +486,29 @@ jl_clean(){
 	#rm edit/root/.bash_history # it is convenient to not delete it by default. You should clean up in final build manually.
 	rm edit/var/lib/dbus/machine-id
 	rm edit/sbin/initctl
-	msg_out "You have $timeout seconds each to answere the following questions. if not answered, I will take 'n' as default (be ready). Some default may be different due to previous choice.\n"
+	msg_out "You have $timeout seconds each to answer the following questions. if not answered, I will take 'n' as default (be ready). Some default may be different due to previous choice.\n"
+	homec=$(get_prop_yn "$JL_rhpn" "$liveconfigfile" "Retain home directory" "$timeout")
+	if [  "$homec" = Y ] || [ "$homec" = y ]; then
+	  	msg_out "edit/home kept as it is"
+	else
+	  	rm -rf edit/home/*
+	  	msg_out "edit/home cleaned!"
+	fi
+	update_prop_val "$JL_rhpn" "$homec"  "$liveconfigfile" "Whether to keep users home directory, by default it is deleted."
+}
+
+jl_clean_arch(){
+	kerver=$(uname -r)
+	cd "$livedir" #exported from jlcd_start
+	if [ -d edit/mydir ]; then
+		mv -f edit/mydir ./
+	fi
+	rm -rf edit/tmp/*
+	rm edit/var/lib/dbus/machine-id
+	rm edit/sbin/initctl
+	$CHROOT $SHELL -c "LANG=C pacman -Sl | awk '/\[installed\]$/ {print \$1 \"/\" \$2 \"-\" \$3}' > /pkglist.txt"
+	mv edit/pkglist.txt extracted/arch/pkglist.$JL_arch.txt
+	msg_out "You have $timeout seconds each to answer the following questions. if not answered, I will take 'n' as default (be ready). Some default may be different due to previous choice.\n"
 	homec=$(get_prop_yn "$JL_rhpn" "$liveconfigfile" "Retain home directory" "$timeout")
 	if [  "$homec" = Y ] || [ "$homec" = y ]; then
 	  	msg_out "edit/home kept as it is"
@@ -475,7 +523,9 @@ jlcd_start(){
 	export livedir=
 	export liveconfigfile=
 	export edit=
-	if $JL_debian; then
+    if $JL_archlinux; then
+        msg_out "Running in Arch Linux mode"
+	elif $JL_debian; then
 		msg_out "Running in Debian mode"
 	else
 		msg_out "Running in Ubuntu mode"
@@ -518,6 +568,7 @@ jlcd_start(){
 		c=2
 		cd "$livedir"
 		isopath="$(cat "$JLIVEisopathF")"
+        IMAGENAME="$(isoinfo -d -i "$isopath" |sed -n 's/^[[:blank:]]*volume id:[[:blank:]]*\(.*\)$/\1/ip')"
 		if [ -d edit ]; then
 			wrn_out "seems this isn't really a new project (edit directory exists),existing files will be overwritten!!! if you aren't sure what this warning is about, close this terminal and run again. If this is shown again, enter y and continue..."
 			cont=$(get_yn "Are you sure, you want to continue (y/n)?: " $timeout)
@@ -529,11 +580,11 @@ jlcd_start(){
 			fi
 		fi
 		mount -o loop "$isopath" mnt || wrn_out "failed to mount iso."
-		rsync --exclude=/"$JL_casper"/filesystem.squashfs -a mnt/ extracted || err_exit "rsync failed"
-		unsquashfs mnt/"$JL_casper"/filesystem.squashfs || err_exit "unsquashfs failed"
+        rsync --exclude=/"$JL_squashfs" -a mnt/ extracted || err_exit "rsync failed"
+        unsquashfs mnt/"$JL_squashfs" || err_exit "unsquashfs failed"
 		mv squashfs-root edit || err_exit "couldn't move squashfs-root."
 		edit=$(abs_path edit)/ #must end with a slash
-		umount mnt
+		umount mnt || umount -lf mnt
 	fi
 	cd "$maindir"
 	c=1
@@ -570,6 +621,10 @@ jlcd_start(){
 	touch "$liveconfigfile"
 	chmod 777 "$liveconfigfile"
 	edit=$(abs_path "$livedir/edit")/ #must end with a slash
+    if [ "$IMAGENAME" = '' ]; then
+        IMAGENAME="$(get_prop_val $JL_inpn "$liveconfigfile")"
+    fi
+    update_prop_val "$JL_inpn" "$IMAGENAME" "$liveconfigfile" "Do not modify if you don't know what you are doing."
 
 	set -a
 	if [ -f "$livedir/$JL_sconf"  ]; then
@@ -608,14 +663,23 @@ jlcd_start(){
 	cd "$livedir"
 	msg_out "Entered into directory $livedir"
 	##############################Enable network connection####################################################################
-	refresh_network
-	##############################Debcache management########################################################################
-	msg_out "Debcache Management starting. Moving deb files to edit/var/cache/apt/archives"
+    cp -L /etc/hosts edit/.
+    cp -L /etc/resolv.conf edit/.
+	#refresh_network
+    #JLopt -rn
+	##############################cache management########################################################################
+	msg_out "Cache Management starting. Moving package files to cache dir"
 	cd "$livedir"
 	if [ -d "debcache" ]; then
-	  echo dummy123456 > debcache/dummy123456.deb
-	  mv -f debcache/*.deb edit/var/cache/apt/archives
-	  msg_out "deb files moved. Debcache Management complete!"
+        if $JL_archlinux; then
+            echo dummy123456 > debcache/dummy123456.pkg.tar.xz
+            mv -f debcache/*.xz edit/var/cache/pacman/pkg/
+            msg_out "pkg files moved. Cache Management complete!"
+        else
+            echo dummy123456 > debcache/dummy123456.deb
+            mv -f debcache/*.deb edit/var/cache/apt/archives
+            msg_out "deb files moved. Cache Management complete!"
+        fi
 	fi
 	#more cache
 	if [ -d mydir ] && [ -d edit ]; then
@@ -642,8 +706,10 @@ jlcd_start(){
 		xhost - && msg_out "access control enabled"
 	fi
 
-	msg_out "installing updarp in chroot ..."
-	cp "$JLdir"/updarp edit/usr/bin/updarp
+    if ! $JL_archlinux; then
+        msg_out "installing updarp in chroot ..."
+        cp "$JLdir"/updarp edit/usr/bin/updarp
+    fi
 
 	mount_fs
 	msg_out "Running chroot terminal... \nWhen you are finished, run: exit or simply close the chroot terminal. run 'cat help' or './help' to get help in chroot terminal."
@@ -660,22 +726,34 @@ jlcd_start(){
 			fi
 		fi
 	fi
-
-	msg_out "removing updarp ..."
-	rm edit/usr/bin/updarp
+    if ! $JL_archlinux; then
+        msg_out "removing updarp ..."
+        rm edit/usr/bin/updarp
+    fi
 	msg_out 'Restoring access control state'
 	xhost $bxhost | sed 's/^/\n*** /' && msg_out "xhost restored to initial state."  #leave this variable unquoted
 	##################################Debcache management############################################################
-	msg_out "Debcache Management starting. Moving .deb files to debcache"
+	msg_out "Cache Management starting. Moving package files to local cache dir"
 	cd "$livedir"
 	if [ ! -d "debcache" ]; then
 	  mkdir debcache
 	fi
-	echo dummy123456 > edit/var/cache/apt/archives/dummy123456.deb
-	mv -f edit/var/cache/apt/archives/*.deb debcache
-	msg_out "deb files moved. Debcache Management complete!"
+    if $JL_archlinux; then
+        echo dummy123456 > edit/var/cache/pacman/pkg/dummy123456.pkg.tar.xz
+        mv -f edit/var/cache/pacman/pkg/*.xz debcache
+        msg_out "pkg files moved. Cache Management complete!"
+        $CHROOT pacman -Scc --noconfirm #cleaning cache
+    else
+        echo dummy123456 > edit/var/cache/apt/archives/dummy123456.deb
+        mv -f edit/var/cache/apt/archives/*.deb debcache
+        msg_out "deb files moved. Cache Management complete!"
+    fi
 	##################################Cleaning...#########################################
-	jl_clean
+    if $JL_archlinux; then
+        jl_clean_arch
+    else
+        jl_clean
+    fi
 	###############################Post Cleaning#####################################################################
 	msg_out "Cleaning system"
 	rm -f edit/prepare
@@ -685,48 +763,55 @@ jlcd_start(){
 	kerver=0
 	d=2
 	ker=""
-	msg_out "##### Init script & Kernel related #####\nRebuild the initrd if you have \n1. changed init scripts or kernel modules\n2. installed new kernel and want to boot that kernel in the live session."
-	ker="$(get_yn "Rebuild initrd: (y/n)?: " $timeout)"
+	msg_out "##### Init script & Kernel related #####\nRebuild the initramfs if you have \n1. changed init scripts or kernel modules\n2. installed new kernel and want to boot that kernel in the live session."
+	ker="$(get_yn "Rebuild initramfs: (y/n)?: " $timeout)"
 	if [ "$ker" = "y" ] || [ "$ker" = "Y" ]; then
-		d=1
-		##################### managing initrd################
-		msg_out "Finding initrd name ..."
-		initrd=$(get_initrd_name "extracted/$JL_casper")
-		if [ "$initrd" = ''  ]; then
-			wrn_out "couldn't dtermine initrd name in: extracted/$JL_casper"
-			initrd="$(get_input "Enter the name of initrd archive: ")"
-		fi
-		msg_out "initrd: $initrd"
-		[ "$initrd" !=  "" ] || err_exit "initrd name can not be empty"
+        if ! $JL_archlinux; then
+            d=1
+            ##################### managing initrd################
+            msg_out "Finding initrd name ..."
+            initrd=$(get_initrd_name "extracted/$JL_casper")
+            if [ "$initrd" = ''  ]; then
+                wrn_out "couldn't dtermine initrd name in: extracted/$JL_casper"
+                initrd="$(get_input "Enter the name of initrd archive: ")"
+            fi
+            msg_out "initrd: $initrd"
+            [ "$initrd" !=  "" ] || err_exit "initrd name can not be empty"
 
-		################# managing vmlinuz ###################
-		msg_out "Finding vmlinuz ..."
-		vmlinuz=$(get_vmlinuz_path "extracted/$JL_casper")
-		if [ "$vmlinuz" = '' ]; then
-			wrn_out "Couldn't find vmlinuz in: extracted/$JL_casper"
-			vmlinuz_name=$(get_input "Enter the name of vmlinuz: ")
-			vmlinuz="extracted/$JL_casper/$vmlinuz_name"
-		fi
-		export vmlinuz_name=$(basename "$vmlinuz")
-		msg_out "vmlinuz: $vmlinuz_name"
-		[ "$vmlinuz_name" != "" ] || err_exit "vmlinuz name can not be empty."
+            ################# managing vmlinuz ###################
+            msg_out "Finding vmlinuz ..."
+            vmlinuz=$(get_vmlinuz_path "extracted/$JL_casper")
+            if [ "$vmlinuz" = '' ]; then
+                wrn_out "Couldn't find vmlinuz in: extracted/$JL_casper"
+                vmlinuz_name=$(get_input "Enter the name of vmlinuz: ")
+                vmlinuz="extracted/$JL_casper/$vmlinuz_name"
+            fi
+            export vmlinuz_name=$(basename "$vmlinuz")
+            msg_out "vmlinuz: $vmlinuz_name"
+            [ "$vmlinuz_name" != "" ] || err_exit "vmlinuz name can not be empty."
+        else
+            rebuild_initramfs
+            d=2
+        fi
 	fi
-	while [ $d -eq 1 ]
-	do
-		kerver="$(get_input "Enter the kernel version (take your time on this one) (n to skip, empty to use `uname -r`): ")"
-		if [ "$kerver" = "n" ] || [ "$kerver" = "N" ]; then
-			break
-		elif [ "$kerver" = "" ]; then
-			kerver=$(uname -r)
-		fi
-		vmlinuz_path=edit/boot/vmlinuz-"$kerver"
-		if [ -f "$vmlinuz_path" ]; then
-			rebuild_initrd "$initrd" "$kerver"
-			d=2
-		else
-			err_out "no such kernel version: $kerver"
-		fi
-	done
+    
+    while [ $d -eq 1 ]
+    do
+        kerver="$(get_input "Enter the kernel version (take your time on this one) (n to skip, empty to use `uname -r`): ")"
+        if [ "$kerver" = "n" ] || [ "$kerver" = "N" ]; then
+            break
+        elif [ "$kerver" = "" ]; then
+            kerver=$(uname -r)
+        fi
+        vmlinuz_path=edit/boot/vmlinuz-"$kerver"
+        if [ -f "$vmlinuz_path" ]; then
+            rebuild_initrd "$initrd" "$kerver"
+            d=2
+        else
+            err_out "no such kernel version: $kerver"
+        fi
+    done
+        
 	fastcomp=$(get_prop_yn "$JL_fcpn" "$liveconfigfile" "Use fast compression (ISO size may become larger)" "$timeout")
 	update_prop_val "$JL_fcpn" "$fastcomp" "$liveconfigfile" "y: Fast compression, larger image size. n: smaller image but slower"
 	#check for uefi
@@ -739,41 +824,52 @@ jlcd_start(){
 	msg_out "Updating some required files..."
 	###############################Create CD/DVD##############################################################################
 	cd "$livedir"
-	chmod +w extracted/"$JL_casper"/filesystem.manifest 2>/dev/null
-	$CHROOT dpkg-query -W --showformat='${Package} ${Version}\n' > extracted/"$JL_casper"/filesystem.manifest
+    if ! $JL_archlinux; then
+        chmod +w extracted/"$JL_casper"/filesystem.manifest 2>/dev/null
+        $CHROOT dpkg-query -W --showformat='${Package} ${Version}\n' > extracted/"$JL_casper"/filesystem.manifest
+    fi
 	#no more CHROOT
 	umount_fs
-	cp extracted/"$JL_casper"/filesystem.manifest extracted/"$JL_casper"/filesystem.manifest-desktop
-	sed -i '/ubiquity/d' extracted/"$JL_casper"/filesystem.manifest-desktop
-	sed -i "/"$JL_casper"/d" extracted/"$JL_casper"/filesystem.manifest-desktop
-	rm -f extracted/"$JL_casper"/filesystem.squashfs
-	msg_out "Deleted old filesystem.squashfs.."
-	msg_out "Rebuilding filesystem.squashfs.."
+    if ! $JL_archlinux; then
+        cp extracted/"$JL_casper"/filesystem.manifest extracted/"$JL_casper"/filesystem.manifest-desktop
+        sed -i '/ubiquity/d' extracted/"$JL_casper"/filesystem.manifest-desktop
+        sed -i "/$JL_casper/d" extracted/"$JL_casper"/filesystem.manifest-desktop
+    fi
+    rm -f extracted/"$JL_squashfs"
+	msg_out "Deleted old squashfs.."
+	msg_out "Rebuilding squashfs.."
 	if [ "$fastcomp" = Y ] || [ "$fastcomp" = y ];then
 	  msg_out "Using fast compression. Size may become larger"
-	  mksquashfs edit extracted/"$JL_casper"/filesystem.squashfs -b 1048576 -e edit/boot
+	  mksquashfs edit extracted/"$JL_squashfs" -b 1048576 -e edit/boot
 	else
 	  msg_out "Using exhaustive compression. Size may become lesser"
-	  #mksquashfs edit extracted/"$JL_casper"/filesystem.squashfs -comp xz
-	  mksquashfs edit extracted/"$JL_casper"/filesystem.squashfs -comp xz -e edit/boot
+	  #mksquashfs edit extracted/"$JL_squashfs" -comp xz
+	  mksquashfs edit extracted/"$JL_squashfs" -comp xz -e edit/boot
 	fi
-	printf $(du -sx --block-size=1 edit | cut -f1) > extracted/"$JL_casper"/filesystem.size
-	cd extracted
-	msg_out "Updating md5sums"
-	if [ -f "MD5SUMS" ]; then
-	  rm MD5SUMS
-	  find -type f -print0 | xargs -0 md5sum | grep -v isolinux/boot.cat | tee MD5SUMS
-	fi
-	if [ -f "md5sum.txt" ]; then
-	  rm md5sum.txt
-	  find -type f -print0 | xargs -0 md5sum | grep -v isolinux/boot.cat | tee md5sum.txt
-	fi
+    if ! $JL_archlinux; then
+        printf $(du -sx --block-size=1 edit | cut -f1) > extracted/"$JL_casper"/filesystem.size
+        cd extracted
+        msg_out "Updating md5sums"
+        if [ -f "MD5SUMS" ]; then
+          rm MD5SUMS
+          find -type f -print0 | xargs -0 md5sum | grep -v isolinux/boot.cat | tee MD5SUMS
+        fi
+        if [ -f "md5sum.txt" ]; then
+          rm md5sum.txt
+          find -type f -print0 | xargs -0 md5sum | grep -v isolinux/boot.cat | tee md5sum.txt
+        fi
+    else
+        cd extracted
+        msg_out "Updating md5sums"
+        md5sum "$JL_squashfs" > "$(dirname "$JL_squashfs")/airootfs.md5"
+    fi
 	msg_out "Creating the image"
+    
 	if [ "$uefi" = Y ] || [ "$uefi" = y ];then
-		genisoimage -U -A "$IMAGE_NAME" -V "$IMAGE_NAME" -volset "$IMAGE_NAME" -J -joliet-long -r -v -T -o ../"$cdname".iso -b isolinux/isolinux.bin -c isolinux/boot.cat -no-emul-boot -boot-load-size 4 -boot-info-table -eltorito-alt-boot -e boot/grub/efi.img -no-emul-boot . && msg_out 'Prepared UEFI image'
+		genisoimage -U -A "$IMAGENAME" -V "$IMAGENAME" -volset "$IMAGENAME" -J -joliet-long -r -v -T -o ../"$cdname".iso -b isolinux/isolinux.bin -c isolinux/boot.cat -no-emul-boot -boot-load-size 4 -boot-info-table -eltorito-alt-boot -e boot/grub/efi.img -no-emul-boot . && msg_out 'Prepared UEFI image'
 		uefi_opt=--uefi
 	else
-		genisoimage -D -r -V "$IMAGE_NAME" -cache-inodes -J -no-emul-boot -boot-load-size 4 -boot-info-table -l -b isolinux/isolinux.bin -c isolinux/boot.cat -o ../"$cdname".iso .
+		genisoimage -D -r -V "$IMAGENAME" -cache-inodes -J -no-emul-boot -boot-load-size 4 -boot-info-table -l -b isolinux/isolinux.bin -c isolinux/boot.cat -o ../"$cdname".iso .
 		uefi_opt=
 	fi
 	if [ "$nhybrid" != Y ] && [ "$nhybrid" != y ]; then
